requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.32.0
pylast>=5.2.0
langdetect>=1.0.9
numpy>=1.26.0
scikit-learn>=1.4.0
musicbrainzngs>=0.7.1
liblistenbrainz>=0.5.0
lyricsgenius>=3.0.1
